Add core app
- Move home page to core
- Handle signals in core
