+++ ohr/apps/core/apps.py
@@ -1,0 +1,8 @@
+"""Apps module for core app."""
+from django.apps import AppConfig
+
+
+class CoreConfig(AppConfig):
+    """Core app config class."""
+
+    name = "core"
