+++ ohr/apps/core/models.py
@@ -1,0 +1,3 @@
+"""Models module for core app."""
+# pylint: disable=unused-import
+import ohr.apps.users.signals  # noqa
